Issue #46
Separated DatesAndDownloads functionality and the original class still works!
